internal: Add settings cache.
