+++ packages/lumos/src/index.ts
@@ -1,7 +1,8 @@
 import { BeemoConfig, DriverContext, Tool } from '@beemo/core';
 import { DIR_PATTERN_LIST, ESLINT_DIRS } from '@oriflame/lumos-common';
 
-import { LumosSettings, getSettings } from './helpers/getSettings';
+import { getSettings } from './helpers/getSettings';
+import { LumosSettings } from './types';
 
 function hasNoParams(context: DriverContext, name: string): boolean {
   const { params } = context.args;
